codegenApi/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langdetect
